refactor: remove subcomponents folder from Home component folder
